requirements.txt
absl-py==0.10.0
asgiref==3.3.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
detecto==1.2.0
Django==3.1.3
djangorestframework==3.12.2
dlib==19.21.0
einops==0.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
future==0.18.2
gast==0.3.3
gdown==3.11.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lmdb==1.0.0
Markdown==3.3.3
matplotlib==3.3.2
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.4
pdf2image==1.14.0
Pillow==8.0.1
prefetch-generator==1.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.4.1
Shapely==1.7.1
six==1.15.0
sqlparse==0.4.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-metadata==0.25.0
termcolor==1.1.0
tifffile==2020.10.1
torch==1.7.0
torchvision==0.8.1
tqdm==4.51.0
typing-extensions==3.7.4.3
urllib3==1.25.11
vietocr==0.3.2
Werkzeug==1.0.1
wrapt==1.12.1
